# starmap-contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "星图 StarMap API"
  description: "人才能力星云导航系统 - 后端 API 契约"
  version: "1.0.0"
  contact:
    name: "StarMap Team"
  x-summary: "人才技能图谱知识系统全量 API"
servers:
  - url: "http://localhost:8000/api/v1"
    description: "本地开发环境"
  - url: "https://api.starmap.dev/api/v1"
    description: "评审环境"
security:
  - BearerAuth: []
paths:
...
  /datasources/{source_id}:
    get:
      operationId: getDatasourceDetail
      summary: "数据源详情"
      description: "获取指定数据源的详细配置和统计"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "数据源详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceResponse"
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateDatasource
      summary: "更新数据源"
      description: "更新数据源的权威度评分等配置"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataSourceUpdateRequest"
      responses:
        "200":
          description: "更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceResponse"
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /datasources/{source_id}/stats:
    get:
      operationId: getDatasourceStats
      summary: "数据源统计"
      description: "获取数据源日/周/月采集量、质量趋势"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: "数据源统计"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceStatsResponse"
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /datasources/{source_id}/sync:
    post:
      operationId: triggerSourceSync
      summary: "触发单源同步"
      description: "触发指定数据源的同步（创建 source_sync 流水线运行记录）"
      tags: [数据源管理]
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "同步已触发"
          content:
            application/json:
              schema:
                type: object
                required: [run_id, source_name, status, message]
                properties:
                  run_id:
                    type: string
                  source_name:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "404":
          description: "数据源不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────── 闭环验证 ────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from authentication endpoint"
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: "错误描述"
          example: "岗位不存在"
        code:
          type: string
          description: "错误码（可选，部分端点返回）"
          example: "POSITION_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: "错误发生时间（可选）"

    # ──────────────────────── 数据大屏 ────────────────────────
...
    DataSourceResponse:
      type: object
      required: [id, name, source_type, authority_score, status]
      properties:
        id:
          type: string
        name:
          type: string
        source_type:
          type: string
        authority_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        status:
          type: string
        last_crawl_at:
          type: string
          format: date-time
          nullable: true
        total_records:
          type: integer
        valid_records:
          type: integer
        duplicate_rate:
          type: number
          format: float
        avg_quality_score:
          type: number
          format: float
        config:
          type: object
          additionalProperties: true
    DataSourceUpdateRequest:
      type: object
      properties:
        authority_score:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        status:
          type: string
          nullable: true
        config:
          type: object
          nullable: true
    DataSourceStatsResponse:
      type: object
      required: [source_id, source_name, crawl_volume, quality_trend, total_runs, successful_runs, failed_runs, avg_records_per_run]
      properties:
        source_id:
          type: string
        source_name:
          type: string
        crawl_volume:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
        quality_trend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              score:
                type: number
        total_runs:
          type: integer
        successful_runs:
          type: integer
        failed_runs:
          type: integer
        avg_records_per_run:
          type: number

    # ──────────────────────── 闭环验证 ────────────────────────
...